requirements.txt
fastapi
uvicorn
python-dotenv
lyzr
